# dollmerchant/docs/swagger/swagger.yaml
basePath: /dev/mch/v1
host: 192.168.0.167:9555
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 娃娃机商户平台 Api文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 娃娃机商户平台 Api文档
  version: "1.0"
paths:
...
  /room/add:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 房间名称
        in: formData
        name: room_name
        required: true
        type: string
      - description: 设备ID
        in: formData
        name: device_id
        required: true
        type: string
      - description: 房间图片URl
        in: formData
        name: thumbnail
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增房间
      tags:
      - Room
  /room/del:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 房间ID
        in: formData
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除房间
      tags:
      - Room
  /room/info:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 房间ID
        in: query
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取房间
      tags:
      - Room
  /room/list:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 偏移量
        in: query
        name: offset
        required: true
        type: string
      - description: 总条数
        in: query
        name: total_size
        required: true
        type: string
      - description: 一页多少条数据
        in: query
        name: page_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取房间列表
      tags:
      - Room
  /room/upd:
    put:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 房间ID
        in: formData
        name: room_id
        required: true
        type: string
      - description: 房间名称
        in: formData
        name: room_name
        required: true
        type: string
      - description: 设备ID
        in: formData
        name: device_id
        required: true
        type: string
      - description: 房间图片URl
        in: formData
        name: thumbnail
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 更新房间
      tags:
      - Room
  /settle/apply/withdraw:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 账号
        in: formData
        name: account
        required: true
        type: string
      - description: 账号名称
        in: formData
        name: account_name
        required: true
        type: string
      - description: 所属银行
        in: formData
        name: bank
        required: true
        type: string
      - description: 所属支行
        in: formData
        name: sub_bank
        required: true
        type: string
      - description: 分润数据
        in: formData
        name: benefit_data
        required: true
        type: string
      - description: 发票
        in: formData
        name: invoice
        required: true
        type: string
      - description: 提现金额
        in: formData
        name: amount
        required: true
        type: string
      - description: 快递公司
        in: formData
        name: express_company
        required: true
        type: string
      - description: 快递单号
        in: formData
        name: express_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 申请提现
      tags:
      - Settle
  /settle/apply/withdraw/list:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 偏移量
        in: query
        name: offset
        required: true
        type: string
      - description: 总条数
        in: query
        name: total_size
        required: true
        type: string
      - description: 一页多少条数据
        in: query
        name: page_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取申请提现记录列表
      tags:
      - Settle
...
swagger: "2.0"